Improve loading experience in embedded KYC (#10136)
