+++ changelog/update-9821-improve-loading-experience-in-embedded-kyc
@@ -1,0 +1,4 @@
+Significance: minor
+Type: update
+
+Improve loading experience in embedded KYC
