+++ client/onboarding/kyc/index.tsx
@@ -10,9 +10,9 @@ import WooLogo from 'assets/images/woo-logo.svg';
 import Page from 'components/page';
 import { OnboardingContextProvider } from 'onboarding/context';
 import EmbeddedKyc from 'onboarding/steps/embedded-kyc';
-import strings from 'onboarding/strings';
 import { getConnectUrl } from 'utils';
 import { trackKycExit } from 'wcpay/onboarding/tracking';
+import CloseIcon from 'assets/images/icons/close.svg?asset';
 
 const OnboardingKycPage: React.FC = () => {
 	const urlParams = new URLSearchParams( window.location.search );
@@ -52,12 +52,6 @@ const OnboardingKycPage: React.FC = () => {
 		<Page className="wcpay-onboarding-prototype">
 			<OnboardingContextProvider>
 				<div className="stepper__nav">
-					<button
-						type="button"
-						className={ `stepper__nav-button hide` }
-					>
-						{ strings.back }
-					</button>
 					<img
 						src={ WooLogo }
 						alt="Woo"
@@ -68,7 +62,7 @@ const OnboardingKycPage: React.FC = () => {
 						className="stepper__nav-button"
 						onClick={ handleExit }
 					>
-						{ strings.cancel }
+						<img src={ CloseIcon } alt="Close" />
 					</button>
 				</div>
 				<div className="stepper__wrapper">
